Fix test failures on 32-bit systems due to integer overflow. Note that size_t is a 32-bit type on 32-bit Linux systems.

+++ include/schreier-sims.hpp
@@ -368,16 +368,16 @@ namespace libsemigroups {
 
     //! Returns the size of the group represented by this.
     //!
-    //! \returns the size, a value of \c size_t.
+    //! \returns the size, a value of \c uint64_t.
     //!
     //! \par Parameters
     //! (None)
-    size_t size() {
+    uint64_t size() {
       if (empty()) {
         return 1;
       }
       schreier_sims();
-      size_t out = 1;
+      uint64_t out = 1;
       for (index_type i = 0; i < _base_size; i++) {
         out *= _orbits.size(i);
       }
