+++ src/bmat8.cpp
@@ -367,10 +367,10 @@ namespace libsemigroups {
   namespace bmat8_helpers {
     size_t minimum_dim(BMat8 const& x) noexcept {
       size_t i = 0;
-      size_t c = x.to_int();
-      size_t d = x.to_int();
-      size_t y = x.transpose().to_int();
-      size_t z = x.transpose().to_int();
+      uint64_t c = x.to_int();
+      uint64_t d = x.to_int();
+      uint64_t y = x.transpose().to_int();
+      uint64_t z = x.transpose().to_int();
 
       do {
         d = d >> 8;
